Refactor Peep, Place, Player actions hpp=>h/cpp
