+++ src/openrct2-ui/windows/Guest.cpp
@@ -15,7 +15,7 @@
 #include <openrct2/GameState.h>
 #include <openrct2/Input.h>
 #include <openrct2/actions/GuestSetFlagsAction.h>
-#include <openrct2/actions/PeepPickupAction.hpp>
+#include <openrct2/actions/PeepPickupAction.h>
 #include <openrct2/config/Config.h>
 #include <openrct2/localisation/Localisation.h>
 #include <openrct2/management/Marketing.h>
