+++ src/openrct2-ui/windows/Player.cpp
@@ -13,8 +13,8 @@
 #include <openrct2-ui/windows/Window.h>
 #include <openrct2/Game.h>
 #include <openrct2/Input.h>
-#include <openrct2/actions/PlayerKickAction.hpp>
-#include <openrct2/actions/PlayerSetGroupAction.hpp>
+#include <openrct2/actions/PlayerKickAction.h>
+#include <openrct2/actions/PlayerSetGroupAction.h>
 #include <openrct2/config/Config.h>
 #include <openrct2/drawing/Drawing.h>
 #include <openrct2/interface/Colour.h>
