+++ src/openrct2-ui/windows/Staff.cpp
@@ -16,7 +16,7 @@
 #include <openrct2/Context.h>
 #include <openrct2/Game.h>
 #include <openrct2/Input.h>
-#include <openrct2/actions/PeepPickupAction.hpp>
+#include <openrct2/actions/PeepPickupAction.h>
 #include <openrct2/actions/StaffSetCostumeAction.hpp>
 #include <openrct2/actions/StaffSetOrdersAction.hpp>
 #include <openrct2/actions/StaffSetPatrolAreaAction.hpp>
