+++ src/openrct2/actions/GameActionCompat.cpp
@@ -10,8 +10,8 @@
 #include "GameAction.h"
 #include "GuestSetNameAction.h"
 #include "MazeSetTrackAction.h"
-#include "PlaceParkEntranceAction.hpp"
-#include "PlacePeepSpawnAction.hpp"
+#include "PlaceParkEntranceAction.h"
+#include "PlacePeepSpawnAction.h"
 #include "RideCreateAction.hpp"
 #include "RideDemolishAction.hpp"
 #include "RideSetName.hpp"
