+++ src/openrct2/actions/GameActionRegistration.cpp
@@ -45,11 +45,11 @@
 #include "ParkSetParameterAction.h"
 #include "ParkSetResearchFundingAction.h"
 #include "PauseToggleAction.h"
-#include "PeepPickupAction.hpp"
-#include "PlaceParkEntranceAction.hpp"
-#include "PlacePeepSpawnAction.hpp"
-#include "PlayerKickAction.hpp"
-#include "PlayerSetGroupAction.hpp"
+#include "PeepPickupAction.h"
+#include "PlaceParkEntranceAction.h"
+#include "PlacePeepSpawnAction.h"
+#include "PlayerKickAction.h"
+#include "PlayerSetGroupAction.h"
 #include "RideCreateAction.hpp"
 #include "RideDemolishAction.hpp"
 #include "RideEntranceExitPlaceAction.hpp"
