+++ src/openrct2/actions/PlayerSetGroupAction.cpp
@@ -7,44 +7,22 @@
  * OpenRCT2 is licensed under the GNU General Public License version 3.
  *****************************************************************************/
 
-#pragma once
+#include "PlayerSetGroupAction.h"
 
 #include "../network/network.h"
-#include "GameAction.h"
 
-DEFINE_GAME_ACTION(PlayerSetGroupAction, GAME_COMMAND_SET_PLAYER_GROUP, GameActions::Result)
+void PlayerSetGroupAction::Serialise(DataSerialiser& stream)
 {
-private:
-    NetworkPlayerId_t _playerId{ -1 };
-    uint8_t _groupId{ std::numeric_limits<uint8_t>::max() };
+    GameAction::Serialise(stream);
 
-public:
-    PlayerSetGroupAction() = default;
+    stream << DS_TAG(_playerId) << DS_TAG(_groupId);
+}
+GameActions::Result::Ptr PlayerSetGroupAction::Query() const
+{
+    return network_set_player_group(GetPlayer(), _playerId, _groupId, false);
+}
 
-    PlayerSetGroupAction(NetworkPlayerId_t playerId, uint8_t groupId)
-        : _playerId(playerId)
-        , _groupId(groupId)
-    {
-    }
-
-    uint16_t GetActionFlags() const override
-    {
-        return GameAction::GetActionFlags() | GameActions::Flags::AllowWhilePaused;
-    }
-
-    void Serialise(DataSerialiser & stream) override
-    {
-        GameAction::Serialise(stream);
-
-        stream << DS_TAG(_playerId) << DS_TAG(_groupId);
-    }
-    GameActions::Result::Ptr Query() const override
-    {
-        return network_set_player_group(GetPlayer(), _playerId, _groupId, false);
-    }
-
-    GameActions::Result::Ptr Execute() const override
-    {
-        return network_set_player_group(GetPlayer(), _playerId, _groupId, true);
-    }
-};
+GameActions::Result::Ptr PlayerSetGroupAction::Execute() const
+{
+    return network_set_player_group(GetPlayer(), _playerId, _groupId, true);
+}
